refactor(qwen-cloud): split fetcher into API client + parser + orchestrator
The 824-line QwenCloudUsageFetcher mixed dashboard/API URL resolution, host-override normalization, sec_token resolution, request construction, and three endpoint calls plus parsing plus a legacy-shape fallback.

Split into three files:

- QwenCloudUsageFetcher.swift: orchestration only. Owns URL resolution, host overrides, the sec_token -> API calls -> snapshot flow. Down from 824 LOC to ~230 LOC.
- QwenCloudTokenPlanAPIClient.swift: request construction for the three zeldaHttp.apikeyMgr. endpoints (cornerstoneParam envelope, CSRF / cookie / origin / referer headers, percent-encoded form body) and the response-status mapping (200 / 401 / 403 / other).
- QwenCloudUsageParser.swift: current-shape token-plan parsing plus the legacy Alibaba subscription-summary fallback.

The public parseUsageSnapshot(from:now:) entry point stays on QwenCloudUsageFetcher for source compatibility; it delegates to QwenCloudUsageParser.

No behavior change.

Verified:
- swift test --filter AlibabaCodingPlan: 58 tests passed
- swift test --filter AlibabaTokenPlan: 37 tests passed
- swift test --filter QwenCloud: 27 tests passed
- make check: 0 violations
